+++ src/idunno.Bluesky/Actor/BlueskyServer.cs
@@ -225,7 +225,7 @@ namespace idunno.Bluesky
         /// <summary>
         /// Get a list of suggested actors for the authenticator users. The expected use is discovery of accounts to follow during new account onboarding.
         /// </summary>
-        /// <param name="limit">The number of suggested actors to return.</param>
+        /// <param name="limit">The maximum number of suggested actors to return.</param>
         /// <param name="cursor">An optional cursor for pagination.</param>
         /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
         /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>