+++ src/idunno.Bluesky/Bookmarks/BlueskyServer.cs
@@ -1,0 +1,211 @@
+﻿// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.Extensions.Logging;
+
+using idunno.AtProto;
+using idunno.AtProto.Authentication;
+
+using idunno.Bluesky.Bookmarks;
+using idunno.Bluesky.Bookmarks.Model;
+
+namespace idunno.Bluesky
+{
+    public static partial class BlueskyServer
+    {
+        // Bookmarks - https://github.com/bluesky-social/atproto/pull/4163
+
+        private const string CreateBookmarkEndpoint = "/xrpc/app.bsky.bookmark.createBookmark";
+
+        private const string DeleteBookmarkEndpoint = "/xrpc/app.bsky.bookmark.deleteBookmark";
+
+        private const string GetBookmarksEndpoint = "/xrpc/app.bsky.bookmark.getBookmarks";
+
+        /// <summary>
+        /// Creates a bookmark on the specified account from the specified <paramref name="uri"/> and <paramref name="cid"/>.
+        /// </summary>
+        /// <param name="uri">The <see cref="AtUri"/> of the post to bookmark</param>
+        /// <param name="cid">The <see cref="Cid"/> of the post to bookmark</param>
+        /// <param name="service">The <see cref="Uri"/> of the service to add the bookmark to.</param>
+        /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+        /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+        /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+        /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="cid"/>, <paramref name="service"/>,<paramref name="accessCredentials"/> or <paramref name="httpClient"/> are null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="uri"/> does not point to a post</exception>
+        [UnconditionalSuppressMessage(
+            "Trimming",
+            "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+        [UnconditionalSuppressMessage("AOT",
+            "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+        public static async Task<AtProtoHttpResult<EmptyResponse>> CreateBookmark(
+            AtUri uri,
+            Cid cid,
+            Uri service,
+            AccessCredentials accessCredentials,
+            HttpClient httpClient,
+            Action<AtProtoCredential>? onCredentialsUpdated = null,
+            ILoggerFactory? loggerFactory = default,
+            CancellationToken cancellationToken=default)
+        {
+            ArgumentNullException.ThrowIfNull(uri);
+            ArgumentNullException.ThrowIfNull(cid);
+            ArgumentNullException.ThrowIfNull(uri.Collection);
+            ArgumentOutOfRangeException.ThrowIfNotEqual(uri.Collection, CollectionNsid.Post);
+
+            ArgumentNullException.ThrowIfNull(service);
+            ArgumentNullException.ThrowIfNull(accessCredentials);
+            ArgumentNullException.ThrowIfNull(httpClient);
+
+            AtProtoHttpClient<EmptyResponse> client = new(AppViewProxy, loggerFactory);
+
+            CreateBookmarkRequest request = new (uri, cid);
+
+            AtProtoHttpResult<EmptyResponse> response = await client.Post(
+                service,
+                CreateBookmarkEndpoint,
+                request,
+                credentials: accessCredentials,
+                httpClient: httpClient,
+                jsonSerializerOptions: BlueskyJsonSerializerOptions,
+                onCredentialsUpdated: onCredentialsUpdated,
+                cancellationToken: cancellationToken).ConfigureAwait(false);
+
+            return response;
+        }
+
+        /// <summary>
+        /// Deletes a bookmark on the specified account for the specified <paramref name="uri"/>.
+        /// </summary>
+        /// <param name="uri">The <see cref="AtUri"/> of the post to delete the bookmark for.</param>
+        /// <param name="service">The <see cref="Uri"/> of the service to delete the bookmark from.</param>
+        /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+        /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+        /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+        /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/>, <paramref name="service"/>,<paramref name="accessCredentials"/> or <paramref name="httpClient"/> are null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="uri"/> does not point to a post</exception>
+        [UnconditionalSuppressMessage(
+            "Trimming",
+            "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+        [UnconditionalSuppressMessage("AOT",
+            "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+        public static async Task<AtProtoHttpResult<EmptyResponse>> DeleteBookmark(
+            AtUri uri,
+            Uri service,
+            AccessCredentials accessCredentials,
+            HttpClient httpClient,
+            Action<AtProtoCredential>? onCredentialsUpdated = null,
+            ILoggerFactory? loggerFactory = default,
+            CancellationToken cancellationToken = default)
+        {
+            ArgumentNullException.ThrowIfNull(uri);
+            ArgumentNullException.ThrowIfNull(uri.Collection);
+            ArgumentOutOfRangeException.ThrowIfNotEqual(uri.Collection, CollectionNsid.Post);
+
+            ArgumentNullException.ThrowIfNull(service);
+            ArgumentNullException.ThrowIfNull(accessCredentials);
+            ArgumentNullException.ThrowIfNull(httpClient);
+
+            AtProtoHttpClient<EmptyResponse> client = new(AppViewProxy, loggerFactory);
+
+            DeleteBookmarkRequest request = new(uri);
+
+            AtProtoHttpResult<EmptyResponse> response = await client.Post(
+                service,
+                DeleteBookmarkEndpoint,
+                request,
+                credentials: accessCredentials,
+                httpClient: httpClient,
+                jsonSerializerOptions: BlueskyJsonSerializerOptions,
+                onCredentialsUpdated: onCredentialsUpdated,
+                cancellationToken: cancellationToken).ConfigureAwait(false);
+
+            return response;
+        }
+
+        /// <summary>
+        /// Gets a pageable list of the specified user's bookmarks.
+        /// </summary>
+        /// <param name="limit">The maximum number of suggested actors to return.</param>
+        /// <param name="cursor">An optional cursor for pagination.</param>
+        /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
+        /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+        /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+        /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+        /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+        /// <param name="subscribedLabelers">An optional list of <see cref="Did"/>s of labelers to retrieve labels applied to the account.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>The task object representing the asynchronous operation.</returns>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="service"/>, <paramref name="accessCredentials"/> or <paramref name="httpClient"/> is null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt;=0 or &gt;100.</exception>
+        [UnconditionalSuppressMessage(
+            "Trimming",
+            "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+        [UnconditionalSuppressMessage("AOT",
+            "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+            Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+        public static async Task<AtProtoHttpResult<PagedViewReadOnlyCollection<BookmarkView>>> GetBookmarks(
+            string? cursor,
+            int? limit,
+            Uri service,
+            AccessCredentials accessCredentials,
+            HttpClient httpClient,
+            Action<AtProtoCredential>? onCredentialsUpdated = null,
+            ILoggerFactory? loggerFactory = default,
+            IEnumerable<Did>? subscribedLabelers = null,
+            CancellationToken cancellationToken = default)
+        {
+            ArgumentNullException.ThrowIfNull(service);
+            ArgumentNullException.ThrowIfNull(accessCredentials);
+            ArgumentNullException.ThrowIfNull(httpClient);
+
+            int limitValue = limit ?? 50;
+
+            ArgumentOutOfRangeException.ThrowIfNegative(limitValue);
+            ArgumentOutOfRangeException.ThrowIfZero(limitValue);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan(limitValue, Maximum.Bookmarks);
+
+            AtProtoHttpClient<GetBookmarksResponse> request = new(AppViewProxy, loggerFactory);
+
+            AtProtoHttpResult<GetBookmarksResponse> response = await request.Get(
+                service,
+                $"{GetBookmarksEndpoint}?cursor={cursor}&limit={limit}",
+                credentials: accessCredentials,
+                httpClient: httpClient,
+                jsonSerializerOptions: BlueskyJsonSerializerOptions,
+                onCredentialsUpdated: onCredentialsUpdated,
+                subscribedLabelers: subscribedLabelers,
+                cancellationToken: cancellationToken).ConfigureAwait(false);
+
+            if (response.Succeeded)
+            {
+                return new AtProtoHttpResult<PagedViewReadOnlyCollection<BookmarkView>>(
+                    new PagedViewReadOnlyCollection<BookmarkView>(new List<BookmarkView>(response.Result.Bookmarks).AsReadOnly(), response.Result.Cursor),
+                    response.StatusCode,
+                    response.HttpResponseHeaders,
+                    response.AtErrorDetail,
+                    response.RateLimit);
+            }
+            else
+            {
+                return new AtProtoHttpResult<PagedViewReadOnlyCollection<BookmarkView>>(
+                    new PagedViewReadOnlyCollection<BookmarkView>(),
+                    response.StatusCode,
+                    response.HttpResponseHeaders,
+                    response.AtErrorDetail,
+                    response.RateLimit);
+            }
+        }
+    }
+}