+++ src/idunno.Bluesky/Constants.cs
@@ -109,6 +109,11 @@ namespace idunno.Bluesky
         /// The maximum number of messages in a message batch.
         /// </summary>
         public static readonly int BatchedMessages = 100;
+
+        /// <summary>
+        /// The maximum number of bookmarks that can be returned from GetBookmarks.
+        /// </summary>
+        public static readonly int Bookmarks = 100;
     }
 
     /// <summary>